chore: move DropHint from graph to data-display directory
