+++ v3/src/components/data-display/components/drop-hint.scss
