+++ v3/src/components/data-display/components/droppable-svg.tsx
@@ -2,7 +2,7 @@ import { Active, useDroppable } from "@dnd-kit/core"
 import React, { CSSProperties, memo } from "react"
 import { createPortal } from "react-dom"
 import { useOverlayBounds } from "../../../hooks/use-overlay-bounds"
-import { DropHint } from "../../graph/components/drop-hint"
+import { DropHint } from "./drop-hint"
 
 import "./droppable-svg.scss"
