+++ v3/src/components/graph/components/droppable-add-attribute.tsx
@@ -3,9 +3,9 @@ import {clsx} from "clsx"
 import {Active, useDroppable} from "@dnd-kit/core"
 import {useDropHintString} from "../../../hooks/use-drop-hint-string"
 import {getDragAttributeInfo, useDropHandler} from "../../../hooks/use-drag-drop"
-import {DropHint} from "./drop-hint"
 import {PlotType} from "../graphing-types"
 import {GraphPlace} from "../../axis-graph-shared"
+import {DropHint} from "../../data-display/components/drop-hint"
 import {graphPlaceToAttrRole} from "../../data-display/data-display-types"
 import {useGraphDataConfigurationContext} from "../hooks/use-graph-data-configuration-context"
 import {useInstanceIdContext} from "../../../hooks/use-instance-id-context"
