+++ v3/src/components/graph/components/droppable-axis.tsx
@@ -5,7 +5,7 @@ import React, { CSSProperties } from "react"
 import { createPortal } from "react-dom"
 import { AxisPlace } from "../../axis/axis-types"
 import { useGraphLayoutContext } from "../hooks/use-graph-layout-context"
-import { DropHint } from "./drop-hint"
+import { DropHint } from "../../data-display/components/drop-hint"
 
 interface IProps {
   place: AxisPlace
